requirements-dev.txt
-e .[dev]
pytest-sugar>=0.9.7
pytest-randomly>=3.15.0
pytest-benchmark>=4.0.0
pytest-timeout>=2.2.0
pytest-xdist>=3.3.1
black>=23.11.0
mypy>=1.7.0
flake8>=6.1.0 